requirements.txt
accelerate
diffusers
peft
pillow
safetensors
torch
torchvision
transformers
